requirements.txt
streamlit
pandas
openpyxl
plotly
azure-storage-blob
python-dotenv

